Love-8/Love-task8-task-202411081445/gen_test/LibrarySystemTest.java
package test;

import org.junit.Before;
import org.junit.Test;
import static org.junit.Assert.*;

public class LibrarySystemTest {

    @Test
    public void testLoadBooksFromFile() {
        LibrarySystem system = new LibrarySystem();
        ArrayList<Book> books = system.loadBooksFromFile("books.txt");
        assertNotNull(books);
        assertFalse(books.isEmpty());
    }

    @Test
    public void testLoadBooksFromEmptyFile() {
        LibrarySystem system = new LibrarySystem();
        ArrayList<Book> books = system.loadBooksFromFile("empty_books.txt");
        assertNotNull(books);
        assertTrue(books.isEmpty());
    }

    @Test
    public void testLoadBooksFromInvalidFile() {
        LibrarySystem system = new LibrarySystem();
        ArrayList<Book> books = system.loadBooksFromFile("non_existing.txt");
        assertNotNull(books);
        assertTrue(books.isEmpty());
    }
}
```

Please create separate files for each test class and the necessary production code classes. Ensure that test paths to files like "test_cards.txt", "empty.txt", etc. exist in your environment or adjust the paths accordingly.